[Server-Update] ContentSpringDataJpaRepository 202210101552
- change Long to Integer

+++ tannae-server/src/main/java/codesver/tannae/repository/content/ContentSpringDataJpaRepository.java
@@ -3,6 +3,6 @@ package codesver.tannae.repository.content;
 import codesver.tannae.domain.Content;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContentSpringDataJpaRepository extends JpaRepository<Content, Long> {
+public interface ContentSpringDataJpaRepository extends JpaRepository<Content, Integer> {
     
 }
